packages/python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "runtime-fence"
version = "1.0.0"
description = "AI Agent Safety Control - Kill switch and guardrails for AI agents"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Runtime Fence Team", email = "team@runtimefence.com"}
]
keywords = [
    "ai",
    "agent",
    "safety",
    "killswitch",
    "guardrails",
    "security",
    "llm",
    "autogpt",
    "langchain"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
requires-python = ">=3.10"
dependencies = [
    "requests>=2.28.0",
    "psutil>=5.9.0",
    "pystray>=0.19.0",
    "Pillow>=9.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0"
]
alerts = [
    "twilio>=8.0.0"
]
all = [
    "runtime-fence[dev,alerts]"
]

[project.urls]
Homepage = "https://runtimefence.com"
Documentation = "https://github.com/Protocol14019/ai-agent-killswitch/wiki"
Repository = "https://github.com/Protocol14019/ai-agent-killswitch"
Issues = "https://github.com/Protocol14019/ai-agent-killswitch/issues"

[project.scripts]
fence = "cli:main"
runtime-fence = "cli:main"

[project.gui-scripts]
fence-tray = "fence_tray:main"

[tool.setuptools]
py-modules = ["runtime_fence", "cli", "alerts", "agent_scanner", "fence_proxy", "fence_tray", "safe_resume"]

[tool.setuptools.package-data]
runtime_fence = ["*.json", "*.yaml"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
